requirements.txt
apache-airflow-providers-postgres
requests
numpy
onnxruntime
onnxscript
Pillow     
python-dotenv
aiohttp